fix: normalize inventory unit conversion across stock modals
Use one bidirectional base-unit conversion path across stock receipt and recipe modals, preserve decimal quantities in GraphQL and StockItem, and route low-stock receipts through the inbound stock flow. Backend and frontend CI passed.

+++ src/components/Dashboard_Manager/Storage/graphql/inventory.gql.js
@@ -459,7 +459,7 @@ export const ADJUST_STOCK = gql`
     $restaurantId: ID!
     $warehouseId: ID!
     $ingredientId: ID!
-    $qty: Int!
+    $qty: Float!
     $reason: String
   ) {
     adjustStock(
@@ -482,7 +482,7 @@ export const RECEIVE_STOCK = gql`
     $restaurantId: ID!
     $warehouseId: ID!
     $ingredientId: ID!
-    $qty: Int!
+    $qty: Float!
     $costPerBaseUnit: Float!
     $reason: String
     $lot: String
